+++ source/cognito-trigger/setJestEnvironmentVariables.ts
@@ -1,0 +1,5 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+process.env.LOGGING_LEVEL = "VERBOSE";
+process.env.AWS_LAMBDA_FUNCTION_NAME = "fn-name";