+++ source/ui/src/index.js
@@ -3,16 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import "@aws-amplify/ui-react/styles.css";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
+import React from 'react'
+import ReactDOM from 'react-dom'
+import './index.css'
+import App from './App'
+import {UserContextProvider} from './contexts/UserContext'
 
-ReactDOM.render(<App />, document.getElementById("root"));
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+ReactDOM.render(
+    <React.StrictMode>
+        <UserContextProvider>
+            <App />
+        </UserContextProvider>
+    </React.StrictMode>,
+    document.getElementById('root')
+);